fix: app file refactored to ts
