+++ client/src/main.jsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App.jsx';
+import App from './App.tsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
